首页和认证部分后台接口

+++ game-core/game-core-entity/src/main/java/com/fulu/game/core/entity/UserTechAuth.java
@@ -50,7 +50,7 @@ public class UserTechAuth implements Serializable {
     private Integer resourceType;
 
     // 技能综合评分
-    private Integer scoreAvg;
+    private BigDecimal scoreAvg;
 
     private String mobile;
